+++ src/values/processEvent.ts
@@ -1,7 +1,15 @@
-export const ProcessEvent: { [key: string]: string } = {
+export const ProcessEvent: ProcessEvent = {
     STD_OUT: 'out',
     ERR_OUT: 'errorOut',
     ERROR: 'error',
     END: 'end',
     KILLED: 'killed'
+};
+
+interface ProcessEvent {
+    STD_OUT: string;
+    ERR_OUT: string;
+    ERROR: string;
+    END: string;
+    KILLED: string;
 };