+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [2.1.0 - 2021-12-13]
+### Added
+- Added new process to get all files under git control
+- Enhanced values types with interfaces to better use
+
 ## [2.0.0 - 2021-12-13]
 ### Added
 - Changed to Typescript